Fix bug with layout widget removal, refresh word select on note add
When removing widgets from a layout, we were looping on an index while
also popping from the end of the list, so half-way through the loop,
we would get None behavior. Loop backwards and use widget.setParent(None)
over widget.deleteLater().

This was needed to be able to properly refresh the word select pane
on note add. Now users will see a live update of the text when they add
notes with words found, and they are able to click on the now-found words
to see the notes they just created and edit them.

See https://stackoverflow.com/a/25330164/3203662